+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `apiCatalogPriority: "childs"` doesn't hide collection items in the sidebar anymore
+
 ## [5.1.0-rc.1] - 2026-08-15
 
 ### Added